added some examples in comments
Added some examples for Windows Download recipes for future
functionality. I’m just now realizing that adding windows support will
be harder than just the recipe generator because in the current
structure the download recipe has some unique parts to it. I do wonder
if some of that can be moved to the BigFix recipe instead of the
download recipe.