all arguments added for AutoPkgBESEngine processor
All of the arguments have been added for the AutoPkgBESEngine processor.

The ActionScript argument needs to be generalized much better, but some
of the needed items for certain types are there. It will need some
further generalization and it will need some If-Then-Else statements to
handle different kinds of installation types.