Add Initialized preference setting to known keys
Follow up to dfbdc17f473bbd71358e0156a818a3ec00874528: The "Initialized" setting is true after Recipe Robot runs for the first time, and is used to determine whether the configuration sheet should be shown on launch. Adding this key to the known keys prevents the script from removing it from the preference domain and causing the app config sheet to show at every launch.